+++ hnswlib/hnswlib.h
@@ -212,7 +212,6 @@ AlgorithmInterface<dist_t>::searchKnnCloserFirst(const void* query_data, size_t 
 
 #include "space_l2.h"
 #include "space_ip.h"
-#include "space_mv.h"
-#include "epsilon_search.h"
+#include "stop_condition.h"
 #include "bruteforce.h"
 #include "hnswalg.h"
